Update binning information
Binning information is gathered by pattern models. Pattern models, consistent with the approach of returning a pattern in the coordinates of the camera, will return patterns at the same resolution as the camera (not as the camera sensor). The binning info is passed to SLM buffer during AT, anfd the AT is adjusted accordingly.